Extract AWS-specific parts from the common deploy function (#980)
