+++ tasks/aws/ecs.py
@@ -7,7 +7,7 @@ from invoke.tasks import task
 from pydantic import ValidationError
 
 from tasks import config, doc, tool
-from tasks.deploy import deploy
+from tasks.aws.deploy import deploy
 from tasks.destroy import destroy
 
 scenario_name = "aws/ecs"
