+++ tasks/aws/eks.py
@@ -10,7 +10,7 @@ from invoke.tasks import task
 from pydantic import ValidationError
 
 from tasks import config, doc, tool
-from tasks.deploy import deploy
+from tasks.aws.deploy import deploy
 from tasks.destroy import destroy
 
 scenario_name = "aws/eks"
